+++ build-tools/src/main/package/executableTemplates/unicore-uftpd-start.sh
@@ -36,9 +36,10 @@ export PYTHONPATH=$UFTPD_LIB
 
 if [ -e "$SETPRIV" ]
  then
+  export USER_ID=$(id -u $USER)
+  export GROUP_ID=$(id -g $USER)
   echo "Starting as $USER ($USER_ID:$GROUP_ID) with capabilites: $CAPS"
   $SETPRIV --ambient-caps="$CAPS" --inh-caps="$CAPS" --reuid $USER_ID --regid $GROUP_ID --clear-groups  $PYTHON $UFTPD_LIB/UFTPD.py & echo $! > ${UFTPD_PID}
  else
   $PYTHON $UFTPD_LIB/UFTPD.py & echo $! > ${UFTPD_PID}
 fi
-