+++ CHANGES.md
@@ -5,8 +5,13 @@ Changelog for the PyUFTPD server
 
 [Full documentation](https://uftp-docs.readthedocs.io)
 
+UFTPD 3.4.1 (released Apr dd, 2024)
+-----------------------------------
+ - fix: enable SO_KEEPALIVE on data connections as well
+ - fix: UFTPD should still run on (old) Python 3.6.8
+ - minor code cleanup
 
-UFTPD 3.4.0 (released Mar DD, 2024)
+UFTPD 3.4.0 (released Mar 05, 2024)
 -----------------------------------
  - new feature: report MaxSessionsPerClient setting to Auth server in "ping" reply
  - new feature: add PAM.py to optionally put UFTPD processes into the systemd user slice